requirements.txt
attrs==25.4.0
aws-cdk-lib==2.236.0
aws-cdk.asset-awscli-v1==2.2.263
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==48.20.0
boto3==1.42.40
botocore==1.42.40
cattrs==25.3.0
certifi==2026.1.4
charset-normalizer==3.4.4
constructs==10.4.5
idna==3.11
importlib_resources==6.5.2
jmespath==1.1.0
jsii==1.126.0
publication==0.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
typeguard==2.13.3
typing_extensions==4.15.0
urllib3==2.6.3
